feature/#51-특정 물품 상세 조회 기능 구현했습니다

+++ src/main/java/com/example/shop/admin/dto/ProductFilterRequest.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class ProductFilterRequest {
 
 
-    private int minQuantity;
-    private int minPrice;
-    private int maxPrice;
+    private Long minQuantity;
+    private Long minPrice;
+    private Long maxPrice;
 }
